edit order

+++ src/main/java/com/mbicycle/imarket/services/impl/OrderServiceImpl.java
@@ -1,6 +1,7 @@
 package com.mbicycle.imarket.services.impl;
 
 import com.mbicycle.imarket.beans.entities.Order;
+import com.mbicycle.imarket.beans.entities.Profile;
 import com.mbicycle.imarket.daos.OrderRepository;
 import com.mbicycle.imarket.services.interfaces.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -14,18 +15,36 @@ public class OrderServiceImpl implements OrderService {
 
     @Override
     public boolean add(Order order) {
-        orderRepository.save(order);
-        return true;
+        if (findByProfile(order.getProfile()) == null) {
+            orderRepository.save(order);
+        }
+        return findByProfile(order.getProfile()) != null;
     }
 
     @Override
-    public List<Order> getAll(){
+    public List<Order> getAll() {
         return orderRepository.findAll();
     }
 
     @Override
+    public boolean update(Order order) {
+        Order order1 = order;
+        if (orderRepository.findByProfile(order.getProfile()) != null){
+            order1 = orderRepository.update(order);
+        }
+        return !order.equals(order1);
+    }
+
+    @Override
+    public Order findByProfile(Profile profile) {
+        return orderRepository.findByProfile(profile);
+    }
+
+    @Override
     public boolean delete(Order order) {
-        orderRepository.delete(order);
-        return true;
+        if (findByProfile(order.getProfile()) != null) {
+            orderRepository.delete(order);
+        }
+        return findByProfile(order.getProfile()) == null;
     }
 }
